Finalizing the new inventory system design. This will also fix majority of the issues with inventories at the moment due to better core functionality overall.
The new system addresses many issues with the current inventory system in TNE. This new system will also allow us to easily implement trading later on when we decide to do so. The new system adds the ability to blacklist items, slots, and fixes sync issues, and the issue of getting charge when your pin is incorrect and/or account is locked.